Refactor Permutation, PLUQ, and Inverse

+++ test/CuModMatrix/inplace_operations_test.jl
@@ -107,8 +107,8 @@ function test_inplace_basic_operations()
     @test Array(E) == convert.(Float32,expected)
     
     # Test copyto!
-    F_data = convert.(Float64,rand(1:(modulus-1),100,100))
-    F = GPUFiniteFieldMatrices.zeros(Float64,100,100,modulus)
+    F_data = convert.(Float32,rand(1:(modulus-1),100,100))
+    F = GPUFiniteFieldMatrices.zeros(Float32,100,100,modulus)
 
     copyto!(F,F_data)
     @test Array(F) == F_data
